tests: Split up pipeline_shader.cpp tests
